Support regexes in the rules
This seems like the cleanest way to give complete flexibility to the
user. An alternative would be to create an ad hoc wild card system, but
that would make the implementation more complicated, and it would get
worse with more edge cases. At least with regexes, we can always suggest
a regex if the user is not familiar.

We have to remove the "Try It" button because now the rules can have
special characters.